// packages-archived/coil-extension/src/types/commands.ts
export interface Command<T = any> {
  command: string
  data?: T
}

/**
 * content -> background
 * browser.runtime.sendMessage
 */
export interface Logout extends Command {
  command: 'logout'
}

/**
 * content -> background
 * browser.runtime.sendMessage
 */
export interface ContentScriptInit extends Command {
  command: 'contentScriptInit'
  data: {
    origin: string
  }
}

/**
 * content -> background
 * browser.runtime.sendMessage
 */
export interface AdaptedPageDetails extends Command {
  command: 'adaptedPageDetails'
  data: { url: string; channelId?: string }
}
// ...
export interface FetchYoutubeChannelId extends Command {
  command: 'fetchYoutubeChannelId'
  data: {
    youtubeUrl: string
  }
}